Add enemy class
