+++ ProtectTheBody/Assets/Animations/enemyGreenDie.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: enemy2Die
+  m_Name: enemyGreenDie
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
@@ -20,23 +20,23 @@ AnimationClip:
   m_PPtrCurves:
   - curve:
     - time: 0
-      value: {fileID: -7269258962348381619, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: -3467482129086750264, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.083333336
-      value: {fileID: -7538526573375343284, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: -214713014762438076, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.16666667
-      value: {fileID: 7774825562025705289, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 5763219697047046305, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.25
-      value: {fileID: 6896820858711090432, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 3310406727941600490, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.33333334
-      value: {fileID: -3598846908551975315, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 3996227678735594071, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.41666666
-      value: {fileID: 1623746244615523355, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 8123871759619910752, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.5
-      value: {fileID: 9217486657816430990, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 3076809825887105142, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.5833333
-      value: {fileID: 1747797853621192748, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: 770636719131123897, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     - time: 0.6666667
-      value: {fileID: -8878562292591273568, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+      value: {fileID: -786631857748748909, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -56,15 +56,15 @@ AnimationClip:
       customType: 23
       isPPtrCurve: 1
     pptrCurveMapping:
-    - {fileID: -7269258962348381619, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: -7538526573375343284, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: 7774825562025705289, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: 6896820858711090432, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: -3598846908551975315, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: 1623746244615523355, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: 9217486657816430990, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: 1747797853621192748, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
-    - {fileID: -8878562292591273568, guid: 5077d441ed42ed245b16244868967e9b, type: 3}
+    - {fileID: -3467482129086750264, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: -214713014762438076, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 5763219697047046305, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 3310406727941600490, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 3996227678735594071, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 8123871759619910752, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 3076809825887105142, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: 770636719131123897, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
+    - {fileID: -786631857748748909, guid: f2b1518bd7116bf49a25e0f88ce18801, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
