+++ ProtectTheBody/Assets/Animations/enemyGreenDie.anim.meta
