+++ ProtectTheBody/Assets/Animations/enemyGreenIdle.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: enemyIdle
+  m_Name: enemyGreenIdle
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
