+++ ProtectTheBody/Assets/Animations/enemyGreenIdle.anim.meta
