+++ ProtectTheBody/Assets/Animations/enemyRedDie.anim.meta
