+++ ProtectTheBody/Assets/Animations/enemyRedIdle.anim.meta
