+++ ProtectTheBody/Assets/Prefabs/EnemyGreen.prefab
@@ -13,9 +13,9 @@ GameObject:
   - component: {fileID: 2747039243637808373}
   - component: {fileID: 7816018019616332291}
   - component: {fileID: 8810234455647450062}
-  - component: {fileID: 1177781347080558430}
+  - component: {fileID: 6826225340857625035}
   m_Layer: 7
-  m_Name: Enemy2
+  m_Name: EnemyGreen
   m_TagString: Enemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -142,7 +142,7 @@ CircleCollider2D:
   m_Offset: {x: 0.0033367276, y: 0}
   serializedVersion: 2
   m_Radius: 0.1429685
---- !u!114 &1177781347080558430
+--- !u!114 &6826225340857625035
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -151,8 +151,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 4487417111802647479}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 34e9699cc956e3946804b4bcc43316d9, type: 3}
+  m_Script: {fileID: 11500000, guid: 9c066e8dc8e906c47a20079d29fc26ea, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 25
   rewardScore: {fileID: 4759741744518933199, guid: 35a8f654b533cdd48b88ee36f53aad53, type: 3}
+  speed: 0
+  rewardScore: {fileID: 0}
