+++ ProtectTheBody/Assets/Scripts/Enemy.cs
@@ -2,40 +2,35 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Enemy : MonoBehaviour
+public abstract class Enemy : MonoBehaviour
 {
-    private Vector2 direction;
-    private Rigidbody2D body;
-    public int speed;
+    public Rigidbody2D body;
+    public GameObject rewardScore;
     private GameObject manager;
     private Animator anim;
-    public GameObject rewardScore;
-
-    //private float lifetime = 0f;
-    //private bool dead = false;
-    private bool tween = false;
-    
-    // Start is called before the first frame update
-    void Start()
+    public string[] enemyAnimations;
+    public int speed;
+    enum EnemyAnimation
     {
-        body = GetComponent<Rigidbody2D>();
-        direction = -transform.position.normalized;
-        body.AddForce(direction * speed, ForceMode2D.Force);
+        Idle,
+        Die
+    }
+    private bool tween = false;
 
+    public abstract void Setup();
+
+    public void Start()
+    {
+        Setup();
         manager = GameObject.Find("Body");
         anim = gameObject.GetComponent<Animator>();
+        anim.Play(enemyAnimations[(int) EnemyAnimation.Idle]);
     }
 
-    private void Update()
+    private void FixedUpdate()
     {
-        /*lifetime += Time.deltaTime;
-        if (lifetime >= 5 && !dead)
-        {
-            Die();
-            dead = true;
-        }*/
         AnimatorStateInfo animation = anim.GetCurrentAnimatorStateInfo(0);
-        if (animation.IsName("enemyDie") && animation.length <= animation.normalizedTime && !tween)
+        if (animation.IsName(enemyAnimations[(int)EnemyAnimation.Die]) && animation.length <= animation.normalizedTime && !tween)
         {
             LeanTween.alpha(gameObject, 0, .8f);
             tween = true;
@@ -54,7 +49,7 @@ public class Enemy : MonoBehaviour
         score.transform.SetParent(GameObject.Find("Canvas").transform, false);
         score.GetComponent<RewardScore>().Setup(this.transform.position, 10);
 
-        anim.Play("enemyDie");
+        anim.Play(enemyAnimations[(int) EnemyAnimation.Die]);
 
         Destroy(gameObject.GetComponent<CircleCollider2D>());
         Destroy(gameObject.GetComponent<Rigidbody2D>());
