+++ ProtectTheBody/Assets/Scripts/EnemyGreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c066e8dc8e906c47a20079d29fc26ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
