+++ ProtectTheBody/Assets/Scripts/EnemyRed.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 34e9699cc956e3946804b4bcc43316d9
+guid: 13f909d2694c25f45aa0a7f5535e681f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
